Add support for namespacing
The XmlDocument constructor now accepts the `xmlns` option, which
is passed to the same option in sax-js.
The XmlElement attribute property will then use a Proxy instead,
in order to keep the old behaviour, but extending it to be
namespace-aware.
All methods that accept a name now have a `NS`-suffixed variant,
which accepts a namespace URI and a local name.
These methods will throw if the `xmlns` option is not enabled.